component file restructuring
