+++ src/components/Windows/AboutMe/AboutMeScreen.tsx
@@ -1,0 +1,10 @@
+import { Window } from "../window";
+// import styles from "AboutMe.module.scss";
+
+export function AboutMeScreen() {
+  return (
+    <Window name="Sobre o Matheus" windowIndex={0}>
+      <h1>Sobre o Matheus!</h1>
+    </Window>
+  );
+}
