+++ src/components/Windows/TalkToMe/TalkToMeScreen.tsx
@@ -1,0 +1,10 @@
+import { Window } from "../window";
+import styles from "./TalkToMe.module.scss";
+
+export function TalkToMeScreen() {
+  return (
+    <Window name="Fale comigo" windowIndex={2}>
+      <div></div>
+    </Window>
+  );
+}
